Update tests for new api for building local data
